fix(mcp): report the voice that spoke, not the one that was asked for
`synthesize_speech` returned `voice: "(auto)"` whenever the caller omitted
`voice` — not a Voice id, so it could not be fed back, cached on, or told to
a user. Worse, the auto-routing it claimed never happened: routing lived in
`src/cli/say.ts`, which the MCP server does not go through, so every voiceless
MCP call synthesized with the engine's English default regardless of language.

`resolveSayVoice` moves to `src/voice-routing.ts` and both entry points call
it. The MCP tool resolves before spawning and passes the id explicitly, so the
id it reports is the one the engine was given. When routing declines — no
detection, or a language this build has no voice for — it names
`DEFAULT_VOICE_ID`, the same voice `say` falls back to on a bare invocation; a
test parses the constant out of rust/src/tts/voices.rs so the two cannot drift.

Closes #942
